feat(react): merge createBundle with createRoomBundle
