+++ packages/react/src/index.ts
@@ -2,11 +2,4 @@ export { ConnectionState, createClient } from "@pluv/client";
 export type { CrdtType, InferCrdtJson } from "@pluv/crdt";
 export type { BaseUser, EventMessage, EventRecord, IOEventMessage } from "@pluv/types";
 export { createBundle } from "./createBundle";
-export type { CreateBundle, PluvProviderProps } from "./createBundle";
-export type {
-    CreateRoomBundle,
-    CreateRoomBundleOptions,
-    InferRoomStorage,
-    PluvRoomProviderProps,
-    SubscriptionHookOptions,
-} from "./createRoomBundle";
+export type { CreateBundle, PluvProviderProps, PluvRoomProviderProps, SubscriptionHookOptions } from "./createBundle";
